refactor(api): split period controllers into two files: 'admin' and 'core'
