+++ packages/api/src/period/controllers/core.ts
@@ -1,0 +1,153 @@
+import { BadRequestError, NotFoundError } from '@error/errors';
+import { getQueryInput, getQuerySort } from '@shared/functions';
+import {
+  PaginatedResponseBody,
+  QueryInputParsedQs,
+  TypedRequestQuery,
+  TypedResponse,
+} from '@shared/types';
+import { praiseWithScore } from '@praise/utils';
+import { PraiseDetailsDto, PraiseDto } from '@praise/types';
+import { praiseDocumentListTransformer } from '@praise/transformers';
+import { PraiseModel } from '@praise/entities';
+import { Request } from 'express';
+import { StatusCodes } from 'http-status-codes';
+import mongoose from 'mongoose';
+import { PeriodModel } from '../entities';
+import { periodDocumentTransformer } from '../transformers';
+import {
+  PeriodDetailsDto,
+  PeriodQuantifierPraiseInput,
+  PeriodReceiverPraiseInput,
+  PeriodStatusType,
+} from '../types';
+import { findPeriodDetailsDto, getPreviousPeriodEndDate } from '../utils';
+
+/**
+ * Description
+ * @param
+ */
+export const all = async (
+  req: TypedRequestQuery<QueryInputParsedQs>,
+  res: TypedResponse<PaginatedResponseBody<PeriodDetailsDto | undefined>>
+): Promise<void> => {
+  const query = getQueryInput(req.query);
+
+  const response = await PeriodModel.paginate({
+    ...query,
+    sort: getQuerySort(req.query),
+  });
+
+  const periodList = response?.docs;
+  if (periodList && Array.isArray(periodList) && periodList.length > 0) {
+    const periodDetailsList: PeriodDetailsDto[] = [];
+    for (const period of periodList) {
+      if (period?.status === PeriodStatusType.QUANTIFY) {
+        const periodDetails = await findPeriodDetailsDto(period._id);
+        periodDetailsList.push(periodDetails);
+        continue;
+      }
+      periodDetailsList.push(periodDocumentTransformer(period));
+    }
+    res.status(StatusCodes.OK).json({
+      ...response,
+      docs: periodDetailsList,
+    });
+  } else {
+    res.status(StatusCodes.OK).json({
+      ...response,
+      docs: [],
+    });
+  }
+};
+
+/**
+ * Description
+ * @param
+ */
+export const single = async (
+  req: Request,
+  res: TypedResponse<PeriodDetailsDto>
+): Promise<void> => {
+  const { periodId } = req.params;
+  const periodDetailsDto = await findPeriodDetailsDto(periodId);
+  res.status(StatusCodes.OK).json(periodDetailsDto);
+};
+
+/**
+ * Description
+ * @param
+ */
+export const receiverPraise = async (
+  req: TypedRequestQuery<PeriodReceiverPraiseInput>,
+  // req: Request,
+  res: TypedResponse<PraiseDto[]>
+): Promise<void> => {
+  const period = await PeriodModel.findById(req.params.periodId);
+  if (!period) throw new NotFoundError('Period');
+
+  const { receiverId } = req.query;
+  if (!receiverId) throw new BadRequestError('Receiver Id is a required field');
+
+  const previousPeriodEndDate = await getPreviousPeriodEndDate(period);
+
+  const praiseList = await PraiseModel.find()
+    .where({
+      createdAt: { $gt: previousPeriodEndDate, $lte: period.endDate },
+      receiver: new mongoose.Types.ObjectId(receiverId),
+    })
+    .sort({ createdAt: -1 })
+    .populate('receiver giver');
+
+  const praiseDetailsDtoList: PraiseDetailsDto[] = [];
+  if (praiseList) {
+    for (const praise of praiseList) {
+      praiseDetailsDtoList.push(await praiseWithScore(praise));
+    }
+  }
+
+  res.status(StatusCodes.OK).json(praiseDetailsDtoList);
+};
+
+/**
+ * Description
+ * @param
+ */
+export const quantifierPraise = async (
+  req: TypedRequestQuery<PeriodQuantifierPraiseInput>,
+  // req: Request,
+  res: TypedResponse<PraiseDto[]>
+): Promise<void> => {
+  const period = await PeriodModel.findById(req.params.periodId);
+  if (!period) throw new NotFoundError('Period');
+
+  const { quantifierId } = req.query;
+  if (!quantifierId)
+    throw new BadRequestError('Quantifier Id is a required field');
+
+  const previousPeriodEndDate = await getPreviousPeriodEndDate(period);
+
+  const praiseList = await PraiseModel.aggregate([
+    {
+      $match: {
+        createdAt: { $gt: previousPeriodEndDate, $lte: period.endDate },
+      },
+    },
+    { $unwind: '$quantifications' },
+    {
+      $match: {
+        $expr: {
+          $eq: [
+            '$quantifications.quantifier',
+            new mongoose.Types.ObjectId(quantifierId),
+          ],
+        },
+      },
+    },
+  ]);
+
+  await PraiseModel.populate(praiseList, { path: 'receiver giver' });
+
+  const response = await praiseDocumentListTransformer(praiseList);
+  res.status(StatusCodes.OK).json(response);
+};
